openxt_setup.sh: Add parameters to name build.
There's some magic in the OXT build scripts that assumes there's a
tag in the git repos if you set all of these new build attributes
to name the build. We cope with this by setting ALLOW_SWITCH_BRANCH_FAIL
which causes the build to fall back to 'master' when the tag it expects
isn't there.

There's a better solution to this so we'll call it a workaround for
now.